subtitles handlers configured as beans and some small code refactorings
