+++ annotation-definitions/pom.xml
@@ -34,6 +34,11 @@
 			<artifactId>commons-definitions</artifactId>
 			<version>${version.commonsApi}</version>
 		</dependency>
+		<dependency>
+		    <groupId>com.fasterxml.jackson.dataformat</groupId>
+		    <artifactId>jackson-dataformat-xml</artifactId>
+		    <version>${version.jackson}</version>
+		</dependency>
 	</dependencies>
 
 	<build>
