+++ annotation-dereferenciation/src/main/resources/deref2json.xsl
