+++ annotation-utils/pom.xml
@@ -101,18 +101,6 @@
 			<artifactId>javax.annotation-api</artifactId>
 			<version>${version.javax.annotation}</version>
 		</dependency>
-		
-		<dependency>
-		    <groupId>com.fasterxml.jackson.dataformat</groupId>
-		    <artifactId>jackson-dataformat-xml</artifactId>
-		    <version>${version.jackson}</version>
-		</dependency>	
-		
-		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-context</artifactId>
-			<version>${version.spring}</version>
-		</dependency>
 						
 	</dependencies>
 </project>
