+++ pom.xml
@@ -69,6 +69,7 @@
 		<version.saxonHE>10.1</version.saxonHE>
 		<version.jaxb>2.3.1</version.jaxb>
 		<version.subtitle-converter>1</version.subtitle-converter>
+		<version.maven-surfire-plugin>2.19.1</version.maven-surfire-plugin>
 		
 	</properties>
 	<modules>
@@ -126,7 +127,7 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-surefire-plugin</artifactId>
-			    <version>2.19.1</version>
+			    <version>${version.maven-surfire-plugin}</version>
 				<configuration>
 	    		<testFailureIgnore>true</testFailureIgnore>
 				</configuration>
